+++ packages/react/src/atomic/Input/src/components/InputSearch/inputSearch.types.ts
@@ -3,7 +3,7 @@ import { InputBaseProps } from "../../input.types";
 
 export type InputSearchProperties = Pick<
   InputBaseProps,
-  "appearance" | "disabled"
+  "appearance" | "disabled" | "aiGenerated"
 >;
 
 export type InputSearchBaseProps = InputSearchProperties &