+++ packages/react/src/atomic/Input/src/input.spec.tsx
@@ -90,5 +90,12 @@ describe("GIVEN <Input />", () => {
         screen.getByTestId("input-element-container").getAttribute("class")
       ).toContain("appearance_ai-generative");
     });
+
+    it("THEN should correctly render the aiGenerated prop as ai-generative appearance", () => {
+      makeSut({ aiGenerated: true });
+      expect(
+        screen.getByTestId("input-element-container").getAttribute("class")
+      ).toContain("appearance_ai-generative");
+    });
   });
 });